Move `ForgotPinResult` to the `user.resetpin` package
