+++ app/src/main/java/org/simple/clinic/user/resetpin/ForgotPinResult.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.user
+package org.simple.clinic.user.resetpin
 
 sealed class ForgotPinResult {
   object Success : ForgotPinResult()
